python/requirements.txt
flask==3.1.2
flask-cors==5.0.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
PyPDF2==3.0.1
python-docx==1.1.2
werkzeug==3.1.3
numpy<2.0

